mac_llm/pyproject.toml
[project]
name = "mac-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.61.1",
    "pillow>=11.1.0",
    "pynput>=1.7.7",
    "pyperclip>=1.9.0",
    "pystray>=0.19.5",
    "requests>=2.32.3",
]
